YouTubeVirtualKeyBoard: try fix GSOD if youtube search is not available
This is try to fix https://github.com/Taapat/enigma2-plugin-youtube/issues/35
I have not able to repeat this, but it seems that if youtube search in the web is not available, then the suggestions are delayed and updateSuggestions are called after closing the YouTubeVirtualKeyBoard window.
I hope this could fix this error.