Text cosmetics, thx GWARDAR close #112
Personally, I don't like to change Ok and OK.
I think it looks ugly if next to a button that is written in a normal camel case is a button that has text written only in capital letters.
But I know there is an argument that Ok is not a word.
I use my Latvian translation, where I can change Ok as I wish, so I change Ok to OK if others want this.