YouTubeUi: fix unicode title split for download job
The problem is that the unicode string split on python2 not work correctly if string is not specified as unicode.
On python3 this problem does not exist and the decode works only on bytes objects.
Therefore for python2 converts the string before split to unicode with exceptin for python 3 where such conversion is not required.