Fix ssl error with some (turkish) LC_CTYPE
On OpenATV 7.1 an individual file LC_CTYPE is used for each language. On turkish this cause ssl error unable to find public key parameters.
Thefore add workaround which changes locale LC_CTYPE to default en_US while using the plugin.
In fact this should be fixed in the image, but for now, while there is no clear reason, I will leave this here.