YouTubeVideoUrl: use web response if dash mp4 format is not used
The response of the ios player contains only dash mp4 formats, so it can be used if the user has set not to use dash mp4.
The web format needs to decrypt nsig, which takes time and resources, but after the signature decryption fix, it is no longer throttled and can be used.

Adds additional dash mp4 formats to video formats.

Improve tests to test all players and m3u8 playlist.